[Grupo 8] [Front] [Feat] [Sales Register]

+++ back-end/src/database/migrations/20220601175906-create-sales.js
@@ -26,7 +26,7 @@ module.exports = {
         }
       },
       total_price: {
-        type: Sequelize.FLOAT(9,2),
+        type: Sequelize.DECIMAL(9,2),
         allowNull: false,
       },
       delivery_address: {
